test: 기존 Update 관련 테스트를 QuestionCommonUpdateController로 이동
