+++ application/src/test/java/com/comssa/api/question/common/controller/QuestionCommonUpdateControllerTest.java
@@ -116,4 +116,27 @@ class QuestionCommonUpdateControllerTest extends ControllerTest {
 				)
 			));
 	}
+
+	@Test
+	@DisplayName("삭제")
+	void deleteLicenseQuestion() throws Exception {
+		final String path = baseApiUrl + idUrl;
+		final String document_Name = "성공";
+
+		mockMvc.perform(RestDocumentationRequestBuilders.delete(path))
+			.andExpect(MockMvcResultMatchers.status().isNoContent())
+			.andDo(print())
+			.andDo(MockMvcRestDocumentation.document(
+				document_Name
+			))
+			.andDo(MockMvcRestDocumentationWrapper.document(
+				document_Name,
+				resource(
+					ResourceSnippetParameters.builder()
+						.tag(tag)
+						.description("삭제")
+						.build()
+				)
+			));
+	}
 }
